+++ NEWS.md
@@ -1,5 +1,18 @@
+## CHANGES IN VERSION 0.99.2
+
+### New Features
+
+* New functions with examples/unit tests:
+  - `import_narrowPeak`: Import narrowPeak files, with automated header
+  annotation using metadata from ENCODE.  
+  - `gather_files`: Automatically peak/picard/bed files and read them in
+  as a list of `GRanges` objects.  
+  - `write_example_peaks`: Write example peak data to disk.
+* Update *.gitignore*
+* Update *.Rbuildignore*
+
 ## CHANGES IN VERSION 0.99.0
 
 ### New Features
 
-*   EpiCompare released to Bioconductor.
+*   `EpiCompare` released to Bioconductor.